requirements.txt
## Software Requirements


Python Version: 3.8
NumPy Version: 1.22.4
Pandas Version: 1.2.4
Scikit-Learn Version: 1.3.2
PyTorch：1.10.0
PyTorch-Lightning：1.6.0
Tslearn：0.6.0
Scikit-Learn-Extra：0.2.0
Matplotlib：3.3.0
Seaborn：0.11.0
Scipy：1.7.0